app/views.py
from rest_framework import viewsets

from app.models import Lead
from app.serializers import LeadSerializer


class LeadViewSet(viewsets.ModelViewSet):
    queryset = Lead.objects.all()
    serializer_class = LeadSerializer
