requirements.txt
Django==4.2.6
psycopg2-binary==2.9.9
django-environ==0.11.2
djangorestframework==3.14
django-filter==23.3
markdown==3.5
pyyaml==6.0.1
uritemplate==4.1.1